pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "colorama>=0.4.6",
    "lxml>=5.3.1",
    "pydantic-settings>=2.8.1",
    "pymongo>=4.10.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
